Pull out group filters & streams into separate component
